Converted the transactions state into a Zustand store
