+++ src/hooks/index.js
@@ -1,12 +1,12 @@
 import useAddItem from "./useAddItem.js";
-import useGetItems from "./useGetItems.js";
+import useTransactionsStore from "./useTransactionsStore.js";
 import useTotalBalanceStore from "./useTotalBalanceStore.js";
 import fetchTransactions from "./fetchTransactions.js";
 import useModalStore from "./useModalStore.js";
 
 export {
   useAddItem,
-  useGetItems,
+  useTransactionsStore,
   useTotalBalanceStore,
   fetchTransactions,
   useModalStore,
